Refactor according to Bills suggestions
