+++ src/ubkg_api/common_routes/codes/codes_controller.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, jsonify, current_app
+from ..neo4j_logic import codes_code_id_codes_get_logic, codes_code_id_concepts_get_logic
 
 codes_blueprint = Blueprint('codes', __name__, url_prefix='/codes')
 
@@ -7,7 +8,6 @@ codes_blueprint = Blueprint('codes', __name__, url_prefix='/codes')
 def codes_code_id_codes_get(code_id, sab=[]):
     """Returns a list of code_ids {Concept, Code, SAB} that code the same concept(s) as the code_id, optionally restricted to source (SAB)
 
-
     :param code_id: The code identifier
     :type code_id: str
     :param sab: One or more sources (SABs) to return
@@ -15,17 +15,18 @@ def codes_code_id_codes_get(code_id, sab=[]):
 
     :rtype: Union[List[CodesCodesObj], Tuple[List[CodesCodesObj], int], Tuple[List[CodesCodesObj], int, Dict[str, str]]
     """
-    return jsonify(current_app.neo4jManager.codes_code_id_codes_get(code_id, sab))
+    neo4j_instance = current_app.neo4jConnectionHelper.instance()
+    return jsonify(codes_code_id_codes_get_logic(neo4j_instance, code_id, sab))
 
 
 @codes_blueprint.route('/<code_id>/concepts', methods=['GET'])
 def codes_code_id_concepts_get(code_id):
     """Returns a list of concepts {Concept, Prefterm} that the code_id codes
 
-
     :param code_id: The code identifier
     :type code_id: str
 
     :rtype: Union[List[ConceptDetail], Tuple[List[ConceptDetail], int], Tuple[List[ConceptDetail], int, Dict[str, str]]
     """
-    return jsonify(current_app.neo4jManager.codes_code_id_concepts_get(code_id))
+    neo4j_instance = current_app.neo4jConnectionHelper.instance()
+    return jsonify(codes_code_id_concepts_get_logic(neo4j_instance, code_id))
