+++ src/ubkg_api/common_routes/tui/tui_controller.py
@@ -1,4 +1,5 @@
 from flask import Blueprint, jsonify, current_app
+from ..neo4j_logic import tui_tui_id_semantics_get_logic
 
 tui_blueprint = Blueprint('tui', __name__, url_prefix='/tui')
 
@@ -7,10 +8,10 @@ tui_blueprint = Blueprint('tui', __name__, url_prefix='/tui')
 def tui_tui_id_semantics_get(tui_id):
     """Returns a list of symantic_types {semantic, STN} of the type unique id (tui)
 
-
     :param tui_id: The TUI identifier
     :type tui_id: str
 
     :rtype: Union[List[SemanticStn], Tuple[List[SemanticStn], int], Tuple[List[SemanticStn], int, Dict[str, str]]
     """
-    return jsonify(current_app.neo4jManager.tui_tui_id_semantics_get(tui_id))
+    neo4j_instance = current_app.neo4jConnectionHelper.instance()
+    return jsonify(tui_tui_id_semantics_get_logic(neo4j_instance, tui_id))
