--- src/ubkg_api/instance/app.cfg
@@ -1,3 +1,0 @@
-SERVER='bolt://18.205.215.12:7688'
-USERNAME='neo4j'
-PASSWORD='HappyG0at'
